refactor: split schedule card
