+++ spec/components/schedules/card/dose_status_component_spec.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Components::Schedules::Card::DoseStatusComponent, type: :component do
+  let(:person) { create(:person) }
+  let(:medication) { create(:medication, name: 'Ibuprofen', current_supply: 1000, supply_at_last_restock: 1000) }
+  let(:schedule) do
+    Schedule.create!(
+      person: person,
+      medication: medication,
+      dose_amount: 400.0,
+      dose_unit: 'mg',
+      frequency: 'Twice daily',
+      max_daily_doses: 4,
+      notes: 'Take with food',
+      start_date: 1.month.ago,
+      end_date: 1.month.from_now
+    )
+  end
+  let(:presenter) { Schedules::CardPresenter.new(schedule: schedule, todays_takes: [take], current_user: nil, person: person) }
+  let(:take) { create(:medication_take, :for_schedule, schedule: schedule, taken_at: Time.current, amount_ml: 400) }
+
+  it 'renders dates, notes, dose count, and today take history' do
+    rendered = render_inline(described_class.new(schedule: schedule, presenter: presenter))
+
+    expect(rendered.text).to include('Take with food')
+    expect(rendered.text).to include("Today's Doses")
+    expect(rendered.text).to include('1/4')
+    expect(rendered.text).to include('400mg')
+  end
+end
